# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "S360°",
        "description": "Documentation interactive de S360°, couvrant les modules migrés depuis S360.",
        "contact": {
            "name": "Équipe S360°",
            "email": "support@s360.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Instance locale de developpement"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/banque/tvf-comp-1/{id}": {
            "get": {
                "tags": [
                    "Banque TVF Comp 1"
                ],
                "summary": "Afficher un complement",
                "operationId": "bdefdccb46d387ae408b18fae9486076",
                "responses": {
                    "200": {
                        "description": "Details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Banque TVF Comp 1"
                ],
                "summary": "Mettre a jour un complement",
                "operationId": "657b624b5ead2c15897be459c4cef471",
                "responses": {
                    "200": {
                        "description": "Mis a jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Banque TVF Comp 1"
                ],
                "summary": "Supprimer un complement",
                "operationId": "75e53de4bae28499417034c2ef1ada4b",
                "responses": {
                    "204": {
                        "description": "Supprime"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Utiliser un token d'accès généré via Sanctum.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}